# docs/openapi-v1.yaml
openapi: 3.1.0
info:
  title: JARVIS Backend API
  version: 1.1.0
  description: |
    v1 API contract aligned with currently implemented Fastify routes.
servers:
  - url: https://api.jarvis.local
    description: Default server
...
paths:
...
  /api/v1/councils/runs:
    post:
      tags: [Councils]
      summary: Start asynchronous council run for deep multi-agent deliberation
      operationId: startCouncilRun
      parameters:
        - in: header
          name: idempotency-key
          required: true
          schema:
            type: string
            minLength: 8
            maxLength: 200
        - in: header
          name: x-trace-id
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CouncilRunCreateRequest'
      responses:
        '202':
          description: Council run accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessCouncilRun'
        '200':
          description: Existing council run returned for idempotency replay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessCouncilRun'
        default:
          $ref: '#/components/responses/ApiError'
    get:
      tags: [Councils]
      summary: List recent council runs
      operationId: listCouncilRuns
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Council run list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessCouncilRuns'
        default:
          $ref: '#/components/responses/ApiError'
  /api/v1/councils/runs/{runId}:
    get:
      tags: [Councils]
      summary: Get single council run
      operationId: getCouncilRun
      parameters:
        - in: path
          name: runId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Council run detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessCouncilRun'
        '404':
          $ref: '#/components/responses/ApiError'
        default:
          $ref: '#/components/responses/ApiError'
  /api/v1/councils/runs/{runId}/events:
    get:
      tags: [Events]
      summary: Stream council run lifecycle events over SSE
      operationId: streamCouncilRunEvents
      parameters:
        - in: path
          name: runId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'text/event-stream payload (event: stream.open|council.run.updated|council.run.completed|council.run.failed|stream.close)'
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/CouncilRunSseDataEnvelope'
              examples:
                streamOpen:
                  summary: event stream.open
                  value:
                    request_id: req_123
                    run_id: run_123
                runUpdated:
                  summary: event council.run.updated
                  value:
                    run_id: run_123
                    timestamp: '2026-02-23T08:31:05.100Z'
                    data:
                      id: run_123
                      question: Should we migrate service A to Rust?
                      status: running
                      consensus_status: null
                      summary: Council run started.
                      participants: []
                      attempts: []
                      provider: null
                      model: pending
                      used_fallback: false
                      task_id: null
                      created_at: '2026-02-23T08:31:04.000Z'
                      updated_at: '2026-02-23T08:31:05.000Z'
                streamClose:
                  summary: event stream.close
                  value:
                    run_id: run_123
        '404':
          $ref: '#/components/responses/ApiError'
        default:
          $ref: '#/components/responses/ApiError'
...
  /api/v1/tasks:
    post:
      tags: [Tasks]
      summary: Create task
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessTask'
        default:
          $ref: '#/components/responses/ApiError'

    get:
      tags: [Tasks]
      summary: List tasks
      operationId: listTasks
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [queued, running, blocked, retrying, done, failed, cancelled]
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Task list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessTaskList'
        default:
          $ref: '#/components/responses/ApiError'
...
components:
  responses:
    ApiError:
      description: Standard API error envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
  schemas:
    ApiMeta:
      type: object
      additionalProperties: true
    ApiErrorEnvelope:
      type: object
      required: [request_id, error]
      properties:
        request_id:
          type: string
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    ProviderAttempt:
      type: object
      required: [provider, status]
      properties:
        provider:
          type: string
          enum: [openai, gemini, anthropic, local]
        status:
          type: string
          enum: [success, failed, skipped]
        latencyMs:
          type: integer
        error:
          type: string
...
    CouncilRunCreateRequest:
      type: object
      required: [question]
      properties:
        question:
          type: string
          minLength: 1
          maxLength: 4000
        system_prompt:
          type: string
        provider:
          type: string
          enum: [auto, openai, gemini, anthropic, local]
          default: auto
        exclude_providers:
          type: array
          maxItems: 4
          items:
            type: string
            enum: [openai, gemini, anthropic, local]
        strict_provider:
          type: boolean
          default: false
        model:
          type: string
        temperature:
          type: number
          minimum: 0
          maximum: 2
        max_output_tokens:
          type: integer
          minimum: 1
          maximum: 32000
        create_task:
          type: boolean
          default: false
        task_title:
          type: string
          minLength: 1
          maxLength: 200
    CouncilParticipant:
      type: object
      required: [role, provider, status, summary]
      properties:
        role:
          type: string
          enum: [planner, researcher, critic, risk, synthesizer]
        provider:
          type: string
          enum: [openai, gemini, anthropic, local]
          nullable: true
        status:
          type: string
          enum: [success, failed, skipped]
        latency_ms:
          type: integer
        summary:
          type: string
        error:
          type: string
    CouncilRun:
      type: object
      required:
        [
          id,
          question,
          status,
          consensus_status,
          summary,
          participants,
          attempts,
          provider,
          model,
          used_fallback,
          task_id,
          created_at,
          updated_at
        ]
      properties:
        id:
          type: string
        question:
          type: string
        status:
          type: string
          enum: [queued, running, completed, failed]
        consensus_status:
          type: string
          enum: [consensus_reached, contradiction_detected, escalated_to_human]
          nullable: true
        summary:
          type: string
        participants:
          type: array
          items:
            $ref: '#/components/schemas/CouncilParticipant'
        attempts:
          type: array
          items:
            $ref: '#/components/schemas/ProviderAttempt'
        provider:
          type: string
          enum: [openai, gemini, anthropic, local]
          nullable: true
        model:
          type: string
        used_fallback:
          type: boolean
        task_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Task:
      type: object
      required: [id, userId, mode, status, title, input, idempotencyKey, createdAt, updatedAt]
      properties:
        id:
          type: string
        userId:
          type: string
        mode:
          type: string
          enum: [chat, execute, council, code, compute, long_run, high_risk, radar_review, upgrade_execution]
        status:
          type: string
          enum: [queued, running, blocked, retrying, done, failed, cancelled]
        title:
          type: string
        input:
          type: object
          additionalProperties: true
        idempotencyKey:
          type: string
        traceId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TaskCreateRequest:
      type: object
      required: [title]
      properties:
        mode:
          type: string
          enum: [chat, execute, council, code, compute, long_run, high_risk, radar_review, upgrade_execution]
          default: execute
        title:
          type: string
          minLength: 1
          maxLength: 200
        input:
          type: object
          additionalProperties: true
...
    CouncilRunStreamOpen:
      type: object
      required: [request_id, run_id]
      properties:
        request_id:
          type: string
        run_id:
          type: string
    CouncilRunStreamClose:
      type: object
      required: [run_id]
      properties:
        run_id:
          type: string
    CouncilRunStreamUpdate:
      type: object
      required: [run_id, timestamp, data]
      properties:
        run_id:
          type: string
        timestamp:
          type: string
          format: date-time
        data:
          $ref: '#/components/schemas/CouncilRun'
    CouncilRunSseDataEnvelope:
      oneOf:
        - $ref: '#/components/schemas/CouncilRunStreamOpen'
        - $ref: '#/components/schemas/CouncilRunStreamClose'
        - $ref: '#/components/schemas/CouncilRunStreamUpdate'
...
    ApiSuccessCouncilRun:
      type: object
      required: [request_id, data, meta]
      properties:
        request_id:
          type: string
        data:
          $ref: '#/components/schemas/CouncilRun'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    ApiSuccessCouncilRuns:
      type: object
      required: [request_id, data, meta]
      properties:
        request_id:
          type: string
        data:
          type: object
          required: [runs]
          properties:
            runs:
              type: array
              items:
                $ref: '#/components/schemas/CouncilRun'
        meta:
          $ref: '#/components/schemas/ApiMeta'
...
    ApiSuccessTask:
      type: object
      required: [request_id, data, meta]
      properties:
        request_id:
          type: string
        data:
          $ref: '#/components/schemas/Task'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    ApiSuccessTaskList:
      type: object
      required: [request_id, data, meta]
      properties:
        request_id:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        meta:
          $ref: '#/components/schemas/ApiMeta'
...